Move the module that defines `instancemethod`.
