+++ comtypes/_post_coinit/unknwn.py
@@ -12,7 +12,7 @@ from comtypes._memberspec import ComMemberGenerator, DispMemberGenerator
 from comtypes._memberspec import STDMETHOD, _ComMemberSpec, _DispMemberSpec
 from comtypes._post_coinit import _cominterface_meta_patcher as _meta_patch
 from comtypes._post_coinit.guid import GUID
-from comtypes._py_instance_method import instancemethod
+from comtypes._post_coinit.instancemethod import instancemethod
 
 logger = logging.getLogger(__name__)
